refactor: typization cleanup
